Правда или действие (часть 1). Доработка.
