+++ js/main.js
@@ -1,5 +1,7 @@
 import {createPictures} from './create-pictures.js';
-import {workForm} from './work-form.js';
+import {initUploadForm} from './init-upload-form.js';
+import {initUploadFormValidation} from './validate-upload-form.js';
 
 createPictures();
-workForm();
+initUploadForm();
+initUploadFormValidation();
